+++ sdk/appconfiguration/azure-data-appconfiguration/src/main/java/com/azure/data/appconfiguration/ConfigurationAsyncClient.java
@@ -31,11 +31,11 @@ import com.azure.data.appconfiguration.models.ConfigurationSetting;
 import com.azure.data.appconfiguration.models.ConfigurationSnapshot;
 import com.azure.data.appconfiguration.models.ConfigurationSnapshotStatus;
 import com.azure.data.appconfiguration.models.FeatureFlagConfigurationSetting;
-import com.azure.data.appconfiguration.models.Label;
-import com.azure.data.appconfiguration.models.LabelFields;
-import com.azure.data.appconfiguration.models.LabelSelector;
+import com.azure.data.appconfiguration.models.SettingLabelSelector;
 import com.azure.data.appconfiguration.models.SecretReferenceConfigurationSetting;
 import com.azure.data.appconfiguration.models.SettingFields;
+import com.azure.data.appconfiguration.models.SettingLabel;
+import com.azure.data.appconfiguration.models.SettingLabelFields;
 import com.azure.data.appconfiguration.models.SettingSelector;
 import com.azure.data.appconfiguration.models.SnapshotFields;
 import com.azure.data.appconfiguration.models.SnapshotSelector;
@@ -1409,14 +1409,14 @@ public final class ConfigurationAsyncClient {
     }
 
     /**
-     * Gets a list of labels by given {@link LabelSelector}
+     * Gets a list of labels by given {@link SettingLabelSelector}
      *
      * <p><strong>Code Samples</strong></p>
      *
      * <!-- src_embed com.azure.data.appconfiguration.configurationasyncclient.listLabels -->
      * <pre>
      * String labelNameFilter = &quot;&#123;labelNamePrefix&#125;*&quot;;
-     * client.listLabels&#40;new LabelSelector&#40;&#41;.setNameFilter&#40;labelNameFilter&#41;&#41;
+     * client.listLabels&#40;new SettingLabelSelector&#40;&#41;.setNameFilter&#40;labelNameFilter&#41;&#41;
      *         .subscribe&#40;label -&gt; &#123;
      *             System.out.println&#40;&quot;label name = &quot; + label&#41;;
      *         &#125;&#41;;
@@ -1430,11 +1430,11 @@ public final class ConfigurationAsyncClient {
      * @return a list of labels as paginated response with {@link PagedFlux}.
      */
     @ServiceMethod(returns = ReturnType.COLLECTION)
-    public PagedFlux<Label> listLabels(LabelSelector selector) {
+    public PagedFlux<SettingLabel> listLabels(SettingLabelSelector selector) {
         final String labelNameFilter = selector == null ? null : selector.getNameFilter();
         final String acceptDatetime = selector == null ? null
             : selector.getAcceptDateTime() == null ? null : selector.getAcceptDateTime().toString();
-        final List<LabelFields> labelFields = selector == null ? null : selector.getFields();
+        final List<SettingLabelFields> labelFields = selector == null ? null : selector.getFields();
         return serviceClient.getLabelsAsync(labelNameFilter, null, acceptDatetime, labelFields);
     }