+++ sdk/appconfiguration/azure-data-appconfiguration/src/main/java/com/azure/data/appconfiguration/ConfigurationClient.java
@@ -33,11 +33,11 @@ import com.azure.data.appconfiguration.models.ConfigurationSetting;
 import com.azure.data.appconfiguration.models.ConfigurationSnapshot;
 import com.azure.data.appconfiguration.models.ConfigurationSnapshotStatus;
 import com.azure.data.appconfiguration.models.FeatureFlagConfigurationSetting;
-import com.azure.data.appconfiguration.models.Label;
-import com.azure.data.appconfiguration.models.LabelFields;
-import com.azure.data.appconfiguration.models.LabelSelector;
+import com.azure.data.appconfiguration.models.SettingLabelSelector;
 import com.azure.data.appconfiguration.models.SecretReferenceConfigurationSetting;
 import com.azure.data.appconfiguration.models.SettingFields;
+import com.azure.data.appconfiguration.models.SettingLabel;
+import com.azure.data.appconfiguration.models.SettingLabelFields;
 import com.azure.data.appconfiguration.models.SettingSelector;
 import com.azure.data.appconfiguration.models.SnapshotFields;
 import com.azure.data.appconfiguration.models.SnapshotSelector;
@@ -1503,14 +1503,14 @@ public final class ConfigurationClient {
     }
 
     /**
-     * Gets a list of labels by given {@link LabelSelector}.
+     * Gets a list of labels by given {@link SettingLabelSelector}.
      *
      * <p><strong>Code Samples</strong></p>
      *
      * <!-- src_embed com.azure.data.appconfiguration.configurationclient.listLabels -->
      * <pre>
      * String labelNameFilter = &quot;&#123;labelNamePrefix&#125;*&quot;;
-     * client.listLabels&#40;new LabelSelector&#40;&#41;.setNameFilter&#40;labelNameFilter&#41;&#41;
+     * client.listLabels&#40;new SettingLabelSelector&#40;&#41;.setNameFilter&#40;labelNameFilter&#41;&#41;
      *         .forEach&#40;label -&gt; &#123;
      *             System.out.println&#40;&quot;label name = &quot; + label.getName&#40;&#41;&#41;;
      *         &#125;&#41;;
@@ -1524,12 +1524,12 @@ public final class ConfigurationClient {
      * @return a list of labels as paginated response with {@link PagedIterable}.
      */
     @ServiceMethod(returns = ReturnType.COLLECTION)
-    public PagedIterable<Label> listLabels(LabelSelector selector) {
+    public PagedIterable<SettingLabel> listLabels(SettingLabelSelector selector) {
         return listLabels(selector, Context.NONE);
     }
 
     /**
-     * Gets a list of labels by given {@link LabelSelector}.
+     * Gets a list of labels by given {@link SettingLabelSelector}.
      *
      * <p><strong>Code Samples</strong></p>
      *
@@ -1538,7 +1538,7 @@ public final class ConfigurationClient {
      * String labelNameFilter = &quot;&#123;labelNamePrefix&#125;*&quot;;
      * Context ctx = new Context&#40;key2, value2&#41;;
      *
-     * client.listLabels&#40;new LabelSelector&#40;&#41;.setNameFilter&#40;labelNameFilter&#41;, ctx&#41;
+     * client.listLabels&#40;new SettingLabelSelector&#40;&#41;.setNameFilter&#40;labelNameFilter&#41;, ctx&#41;
      *         .forEach&#40;label -&gt; &#123;
      *             System.out.println&#40;&quot;label name = &quot; + label.getName&#40;&#41;&#41;;
      *         &#125;&#41;;
@@ -1553,11 +1553,11 @@ public final class ConfigurationClient {
      * @return a list of labels as paginated response with {@link PagedIterable}.
      */
     @ServiceMethod(returns = ReturnType.COLLECTION)
-    public PagedIterable<Label> listLabels(LabelSelector selector, Context context) {
+    public PagedIterable<SettingLabel> listLabels(SettingLabelSelector selector, Context context) {
         final String labelNameFilter = selector == null ? null : selector.getNameFilter();
         final String acceptDatetime = selector == null
             ? null : selector.getAcceptDateTime() == null ? null : selector.getAcceptDateTime().toString();
-        final List<LabelFields> labelFields = selector == null ? null : selector.getFields();
+        final List<SettingLabelFields> labelFields = selector == null ? null : selector.getFields();
         return serviceClient.getLabels(labelNameFilter, null, acceptDatetime, labelFields, context);
     }