+++ sdk/appconfiguration/azure-data-appconfiguration/src/main/java/com/azure/data/appconfiguration/models/SettingLabel.java
@@ -12,19 +12,19 @@ import com.azure.json.JsonWriter;
 import java.io.IOException;
 
 /**
- * The Label model.
+ * The SettingLabel model.
  */
 @Immutable
-public final class Label implements JsonSerializable<Label> {
+public final class SettingLabel implements JsonSerializable<SettingLabel> {
     /*
      * The name of the label.
      */
     private String name;
 
     /**
-     * Creates an instance of Label class.
+     * Creates an instance of SettingLabel class.
      */
-    public Label() {
+    public SettingLabel() {
     }
 
     /**
@@ -46,28 +46,28 @@ public final class Label implements JsonSerializable<Label> {
     }
 
     /**
-     * Reads an instance of Label from the JsonReader.
+     * Reads an instance of SettingLabel from the JsonReader.
      * 
      * @param jsonReader The JsonReader being read.
-     * @return An instance of Label if the JsonReader was pointing to an instance of it, or null if it was pointing to
-     * JSON null.
-     * @throws IOException If an error occurs while reading the Label.
+     * @return An instance of SettingLabel if the JsonReader was pointing to an instance of it, or null if it was
+     * pointing to JSON null.
+     * @throws IOException If an error occurs while reading the SettingLabel.
      */
-    public static Label fromJson(JsonReader jsonReader) throws IOException {
+    public static SettingLabel fromJson(JsonReader jsonReader) throws IOException {
         return jsonReader.readObject(reader -> {
-            Label deserializedLabel = new Label();
+            SettingLabel deserializedSettingLabel = new SettingLabel();
             while (reader.nextToken() != JsonToken.END_OBJECT) {
                 String fieldName = reader.getFieldName();
                 reader.nextToken();
 
                 if ("name".equals(fieldName)) {
-                    deserializedLabel.name = reader.getString();
+                    deserializedSettingLabel.name = reader.getString();
                 } else {
                     reader.skipChildren();
                 }
             }
 
-            return deserializedLabel;
+            return deserializedSettingLabel;
         });
     }
 }