+++ sdk/appconfiguration/azure-data-appconfiguration/src/samples/java/com/azure/data/appconfiguration/ConfigurationClientJavaDocCodeSnippets.java
@@ -14,7 +14,7 @@ import com.azure.core.util.polling.SyncPoller;
 import com.azure.data.appconfiguration.models.ConfigurationSetting;
 import com.azure.data.appconfiguration.models.ConfigurationSettingsFilter;
 import com.azure.data.appconfiguration.models.ConfigurationSnapshot;
-import com.azure.data.appconfiguration.models.LabelSelector;
+import com.azure.data.appconfiguration.models.SettingLabelSelector;
 import com.azure.data.appconfiguration.models.SettingFields;
 import com.azure.data.appconfiguration.models.SettingSelector;
 import com.azure.data.appconfiguration.models.SnapshotFields;
@@ -574,12 +574,12 @@ public final class ConfigurationClientJavaDocCodeSnippets {
     }
 
     /**
-     * Code snippets for {@link ConfigurationClient#listLabels(LabelSelector)}
+     * Code snippets for {@link ConfigurationClient#listLabels(SettingLabelSelector)}
      */
     public void listLabels() {
         // BEGIN: com.azure.data.appconfiguration.configurationclient.listLabels
         String labelNameFilter = "{labelNamePrefix}*";
-        client.listLabels(new LabelSelector().setNameFilter(labelNameFilter))
+        client.listLabels(new SettingLabelSelector().setNameFilter(labelNameFilter))
                 .forEach(label -> {
                     System.out.println("label name = " + label.getName());
                 });
@@ -587,14 +587,14 @@ public final class ConfigurationClientJavaDocCodeSnippets {
     }
 
     /**
-     * Code snippets for {@link ConfigurationClient#listLabels(LabelSelector, Context)}
+     * Code snippets for {@link ConfigurationClient#listLabels(SettingLabelSelector, Context)}
      */
     public void listLabelsMaxOverload() {
         // BEGIN: com.azure.data.appconfiguration.configurationclient.listLabelsMaxOverload
         String labelNameFilter = "{labelNamePrefix}*";
         Context ctx = new Context(key2, value2);
 
-        client.listLabels(new LabelSelector().setNameFilter(labelNameFilter), ctx)
+        client.listLabels(new SettingLabelSelector().setNameFilter(labelNameFilter), ctx)
                 .forEach(label -> {
                     System.out.println("label name = " + label.getName());
                 });