+++ sdk/appconfiguration/azure-data-appconfiguration/src/samples/java/com/azure/data/appconfiguration/ReadmeSamples.java
@@ -21,7 +21,7 @@ import com.azure.data.appconfiguration.models.ConfigurationSettingsFilter;
 import com.azure.data.appconfiguration.models.ConfigurationSnapshot;
 import com.azure.data.appconfiguration.models.FeatureFlagConfigurationSetting;
 import com.azure.data.appconfiguration.models.FeatureFlagFilter;
-import com.azure.data.appconfiguration.models.LabelSelector;
+import com.azure.data.appconfiguration.models.SettingLabelSelector;
 import com.azure.data.appconfiguration.models.SecretReferenceConfigurationSetting;
 import com.azure.data.appconfiguration.models.SettingSelector;
 import com.azure.data.appconfiguration.models.SnapshotSelector;
@@ -472,7 +472,7 @@ public class ReadmeSamples {
     public void listLabels() {
         // BEGIN: readme-sample-listLabels
         String labelNameFilter = "{labelNamePrefix}*";
-        configurationClient.listLabels(new LabelSelector().setNameFilter(labelNameFilter))
+        configurationClient.listLabels(new SettingLabelSelector().setNameFilter(labelNameFilter))
                 .forEach(label -> {
                     System.out.println("label name = " + label.getName());
                 });