+++ sdk/appconfiguration/azure-data-appconfiguration/src/test/java/com/azure/data/appconfiguration/ConfigurationClientTest.java
@@ -20,8 +20,8 @@ import com.azure.data.appconfiguration.models.ConfigurationSetting;
 import com.azure.data.appconfiguration.models.ConfigurationSnapshot;
 import com.azure.data.appconfiguration.models.ConfigurationSnapshotStatus;
 import com.azure.data.appconfiguration.models.FeatureFlagConfigurationSetting;
-import com.azure.data.appconfiguration.models.Label;
-import com.azure.data.appconfiguration.models.LabelSelector;
+import com.azure.data.appconfiguration.models.SettingLabel;
+import com.azure.data.appconfiguration.models.SettingLabelSelector;
 import com.azure.data.appconfiguration.models.SecretReferenceConfigurationSetting;
 import com.azure.data.appconfiguration.models.SettingFields;
 import com.azure.data.appconfiguration.models.SettingSelector;
@@ -1569,17 +1569,17 @@ public class ConfigurationClientTest extends ConfigurationClientTestBase {
 
         // List only the first label var, 'label'
         String label = setting.getLabel();
-        PagedIterable<Label> labels = client.listLabels(new LabelSelector().setNameFilter(setting.getLabel()));
+        PagedIterable<SettingLabel> labels = client.listLabels(new SettingLabelSelector().setNameFilter(setting.getLabel()));
         assertEquals(1, labels.stream().count());
         assertEquals(label, labels.iterator().next().getName());
         // List labels with wildcard label filter
         String label2 = setting2.getLabel();
-        PagedIterable<Label> wildCardLabels = client.listLabels(new LabelSelector().setNameFilter("label*"));
-        List<String> collect = wildCardLabels.stream().collect(Collectors.toList()).stream().map(Label::getName).collect(Collectors.toList());
+        PagedIterable<SettingLabel> wildCardLabels = client.listLabels(new SettingLabelSelector().setNameFilter("label*"));
+        List<String> collect = wildCardLabels.stream().collect(Collectors.toList()).stream().map(SettingLabel::getName).collect(Collectors.toList());
         assertTrue(collect.contains(label));
         assertTrue(collect.contains(label2));
         // List all labels
-        PagedIterable<Label> allLabels = client.listLabels(null);
+        PagedIterable<SettingLabel> allLabels = client.listLabels(null);
         assertTrue(allLabels.stream().count() >= 2);
     }